Extract the import code into the shared module
