+++ crates/completion/Cargo.toml
@@ -15,7 +15,6 @@ log = "0.4.8"
 rustc-hash = "1.1.0"
 either = "1.6.1"
 
-assists = { path = "../assists", version = "0.0.0" }
 stdx = { path = "../stdx", version = "0.0.0" }
 syntax = { path = "../syntax", version = "0.0.0" }
 text_edit = { path = "../text_edit", version = "0.0.0" }
@@ -23,6 +22,7 @@ base_db = { path = "../base_db", version = "0.0.0" }
 ide_db = { path = "../ide_db", version = "0.0.0" }
 profile = { path = "../profile", version = "0.0.0" }
 test_utils = { path = "../test_utils", version = "0.0.0" }
+ide_helpers = { path = "../ide_helpers", version = "0.0.0" }
 
 # completions crate should depend only on the top-level `hir` package. if you need
 # something from some `hir_xxx` subpackage, reexport the API via `hir`.
