+++ crates/completion/src/render.rs
@@ -9,9 +9,9 @@ pub(crate) mod type_alias;
 
 mod builder_ext;
 
-use assists::utils::{ImportScope, MergeBehaviour};
 use hir::{Documentation, HasAttrs, HirDisplay, ModPath, Mutability, ScopeDef, Type};
 use ide_db::RootDatabase;
+use ide_helpers::insert_use::{ImportScope, MergeBehaviour};
 use syntax::TextRange;
 use test_utils::mark;
