+++ crates/ide/Cargo.toml
@@ -24,6 +24,7 @@ stdx = { path = "../stdx", version = "0.0.0" }
 syntax = { path = "../syntax", version = "0.0.0" }
 text_edit = { path = "../text_edit", version = "0.0.0" }
 ide_db = { path = "../ide_db", version = "0.0.0" }
+ide_helpers = { path = "../ide_helpers", version = "0.0.0" }
 cfg = { path = "../cfg", version = "0.0.0" }
 profile = { path = "../profile", version = "0.0.0" }
 test_utils = { path = "../test_utils", version = "0.0.0" }
