+++ crates/rust-analyzer/Cargo.toml
@@ -39,6 +39,7 @@ tracing-tree = { version = "0.1.4" }
 stdx = { path = "../stdx", version = "0.0.0" }
 flycheck = { path = "../flycheck", version = "0.0.0" }
 ide = { path = "../ide", version = "0.0.0" }
+ide_helpers = { path = "../ide_helpers", version = "0.0.0" }
 profile = { path = "../profile", version = "0.0.0" }
 project_model = { path = "../project_model", version = "0.0.0" }
 syntax = { path = "../syntax", version = "0.0.0" }
